Move OMIS params middleware into a shared top level
Sets it up to include other more generic middleware rather than
param specific middleware.
