--- src/apps/omis/middleware/index.js
@@ -1,5 +1,0 @@
-const params = require('./params')
-
-module.exports = {
-  params,
-}
